fix(strategy-audit): port shadow-bundle hardening onto canonical package; drop the diverged shadow copy
The docs/tradeclaw-strategy-audit-patch shadow bundle carried 7 fix commits (2026-06-27/28) that never reached the canonical packages/strategy-audit-agent landed by #134: host:port allowlist entries, www.tradeclaw.win default host, duplicate run-id rejection, signal trade-count parsing, deterministic browser runner hardening (all-request URL gating), and real preset checkbox targeting. Meanwhile #134's landing added identity/payment deny-list hardening the shadow lacked. This union keeps both: shadow behavior everywhere it was the newer lineage, canonical's broader BLOCKED_PATH_PARTS plus its route tests, and new locking tests for host:port allowlisting and duplicate run-id rejection.

Verified: node --test 10/10 pass; npm run check exit 0. Shadow copy under docs/tradeclaw-strategy-audit-patch/packages/ removed; IMPLEMENTATION_README points at the canonical package.
