--- docs/tradeclaw-strategy-audit-patch/packages/strategy-audit-agent/.gitignore
@@ -1,5 +1,0 @@
-node_modules/
-audit-output/
-tmp/
-.env
-.env.*
