--- docs/tradeclaw-strategy-audit-patch/packages/strategy-audit-agent/examples/audit-request.json
@@ -1,20 +1,0 @@
-{
-  "customer": {
-    "name": "Paid Pilot",
-    "reference": "pilot-001"
-  },
-  "objective": "Test whether two standard TradeClaw presets remain usable across BTC and ETH after slippage is enabled.",
-  "strategyRules": "Run the unmodified HMM Top-3 and VWAP + EMA + Bollinger presets. Do not tune parameters after seeing results.",
-  "baseUrl": "https://tradeclaw.win",
-  "symbols": ["BTCUSD", "ETHUSD"],
-  "timeframes": ["H1"],
-  "periods": ["90D"],
-  "strategies": ["HMM Top-3", "VWAP + EMA + Bollinger"],
-  "initialBalance": 10000,
-  "riskPerTradePercent": 1,
-  "includeSlippage": true,
-  "mode": "hybrid",
-  "maxRuns": 8,
-  "allowComputerFallback": true,
-  "outputLabel": "pilot-001"
-}
