--- docs/tradeclaw-strategy-audit-patch/packages/strategy-audit-agent/examples/fixture-results.json
@@ -1,81 +1,0 @@
-[
-  {
-    "id": "btcusd__h1__90d__ema-crossover-rsi",
-    "symbol": "BTCUSD",
-    "timeframe": "H1",
-    "period": "90D",
-    "strategy": "EMA Crossover + RSI",
-    "initialBalance": 10000,
-    "riskPerTradePercent": 1,
-    "includeSlippage": true,
-    "status": "completed",
-    "mode": "deterministic",
-    "metrics": {
-      "totalReturnPercent": 12.4,
-      "netProfit": 1240,
-      "winRatePercent": 44.8,
-      "maxDrawdownPercent": 13.2,
-      "profitFactor": 1.31,
-      "sharpeRatio": 0.88,
-      "expectancy": 24.8,
-      "totalTrades": 50
-    }
-  },
-  {
-    "id": "btcusd__h1__90d__atr-trend-follow",
-    "symbol": "BTCUSD",
-    "timeframe": "H1",
-    "period": "90D",
-    "strategy": "ATR Trend Follow",
-    "initialBalance": 10000,
-    "riskPerTradePercent": 1,
-    "includeSlippage": true,
-    "status": "completed",
-    "mode": "deterministic",
-    "metrics": {
-      "totalReturnPercent": 4.1,
-      "netProfit": 410,
-      "winRatePercent": 39.1,
-      "maxDrawdownPercent": 18.9,
-      "profitFactor": 1.08,
-      "sharpeRatio": 0.31,
-      "expectancy": 8.2,
-      "totalTrades": 50
-    }
-  },
-  {
-    "id": "ethusd__h1__90d__ema-crossover-rsi",
-    "symbol": "ETHUSD",
-    "timeframe": "H1",
-    "period": "90D",
-    "strategy": "EMA Crossover + RSI",
-    "initialBalance": 10000,
-    "riskPerTradePercent": 1,
-    "includeSlippage": true,
-    "status": "completed",
-    "mode": "computer",
-    "metrics": {
-      "totalReturnPercent": -3.2,
-      "netProfit": -320,
-      "winRatePercent": 36.4,
-      "maxDrawdownPercent": 26.3,
-      "profitFactor": 0.92,
-      "sharpeRatio": -0.18,
-      "expectancy": -6.4,
-      "totalTrades": 50
-    }
-  },
-  {
-    "id": "ethusd__h1__90d__atr-trend-follow",
-    "symbol": "ETHUSD",
-    "timeframe": "H1",
-    "period": "90D",
-    "strategy": "ATR Trend Follow",
-    "initialBalance": 10000,
-    "riskPerTradePercent": 1,
-    "includeSlippage": true,
-    "status": "failed",
-    "mode": "hybrid",
-    "error": "Fixture example: result panel did not become visible"
-  }
-]
