--- docs/tradeclaw-strategy-audit-patch/packages/strategy-audit-agent/examples/sample-manifest.json
@@ -1,8 +1,0 @@
-{
-  "schemaVersion": 1,
-  "generatedAt": "2026-06-24T09:38:28.788Z",
-  "requestSha256": "80f5c464595d3322c3edbc0cffe23c6851273daf9a11b7ab64536611a9926485",
-  "resultsSha256": "74c7d27593cfc24c46abe077c86579501ed4812830437bfa04db0e2f6a0b089d",
-  "completedRuns": 3,
-  "totalRuns": 4
-}
